--- Raspite.Example/Raspite.Example.csproj
@@ -1,15 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-	<PropertyGroup>
-		<OutputType>Exe</OutputType>
-		<TargetFramework>net9.0</TargetFramework>
-		<ImplicitUsings>enable</ImplicitUsings>
-		<Nullable>enable</Nullable>
-		<PublishAot>true</PublishAot>
-	</PropertyGroup>
-
-	<ItemGroup>
-		<ProjectReference Include="..\Raspite.Serializer\Raspite.Serializer.csproj"/>
-	</ItemGroup>
-
-</Project>
