--- Raspite.Serializer/Tags/FloatTag.cs
@@ -1,19 +1,0 @@
-﻿namespace Raspite.Serializer.Tags;
-
-public sealed record FloatTag : Tag<float>
-{
-	public override byte Identifier => 5;
-
-	private FloatTag()
-	{
-	}
-
-	public static FloatTag Create(float value, string name = "")
-	{
-		return new FloatTag
-		{
-			Name = name,
-			Value = value
-		};
-	}
-}
