--- Raspite.Tests/Raspite.Tests.csproj
@@ -1,27 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-	<PropertyGroup>
-		<TargetFramework>net9.0</TargetFramework>
-		<LangVersion>latest</LangVersion>
-		<ImplicitUsings>enable</ImplicitUsings>
-		<Nullable>enable</Nullable>
-		<IsPackable>false</IsPackable>
-	</PropertyGroup>
-
-	<ItemGroup>
-		<PackageReference Include="coverlet.collector" Version="6.0.2"/>
-		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.11.1"/>
-		<PackageReference Include="NUnit" Version="4.2.2"/>
-		<PackageReference Include="NUnit.Analyzers" Version="4.3.0"/>
-		<PackageReference Include="NUnit3TestAdapter" Version="4.6.0"/>
-	</ItemGroup>
-
-	<ItemGroup>
-		<Using Include="NUnit.Framework"/>
-	</ItemGroup>
-
-	<ItemGroup>
-		<ProjectReference Include="..\Raspite.Serializer\Raspite.Serializer.csproj"/>
-	</ItemGroup>
-
-</Project>
