+++ Raspite.sln
@@ -3,11 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.4.33122.133
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Raspite.Serializer", "Raspite.Serializer\Raspite.Serializer.csproj", "{F7BFB0F3-E66D-48CF-88B1-4E3A60B2E70A}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Raspite.Example", "Raspite.Example\Raspite.Example.csproj", "{D5EB155C-5365-4E75-935E-CDFA79A9E146}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Raspite.Tests", "Raspite.Tests\Raspite.Tests.csproj", "{D74C9C84-B359-49A5-8B0C-B296C7D8C3D3}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Raspite", "Raspite\Raspite.csproj", "{F7BFB0F3-E66D-48CF-88B1-4E3A60B2E70A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -37,38 +33,6 @@ Global
 		{F7BFB0F3-E66D-48CF-88B1-4E3A60B2E70A}.Release|x64.Build.0 = Release|Any CPU
 		{F7BFB0F3-E66D-48CF-88B1-4E3A60B2E70A}.Release|x86.ActiveCfg = Release|Any CPU
 		{F7BFB0F3-E66D-48CF-88B1-4E3A60B2E70A}.Release|x86.Build.0 = Release|Any CPU
-		{D5EB155C-5365-4E75-935E-CDFA79A9E146}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D5EB155C-5365-4E75-935E-CDFA79A9E146}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D5EB155C-5365-4E75-935E-CDFA79A9E146}.Debug|ARM64.ActiveCfg = Debug|Any CPU
-		{D5EB155C-5365-4E75-935E-CDFA79A9E146}.Debug|ARM64.Build.0 = Debug|Any CPU
-		{D5EB155C-5365-4E75-935E-CDFA79A9E146}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{D5EB155C-5365-4E75-935E-CDFA79A9E146}.Debug|x64.Build.0 = Debug|Any CPU
-		{D5EB155C-5365-4E75-935E-CDFA79A9E146}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{D5EB155C-5365-4E75-935E-CDFA79A9E146}.Debug|x86.Build.0 = Debug|Any CPU
-		{D5EB155C-5365-4E75-935E-CDFA79A9E146}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D5EB155C-5365-4E75-935E-CDFA79A9E146}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D5EB155C-5365-4E75-935E-CDFA79A9E146}.Release|ARM64.ActiveCfg = Release|Any CPU
-		{D5EB155C-5365-4E75-935E-CDFA79A9E146}.Release|ARM64.Build.0 = Release|Any CPU
-		{D5EB155C-5365-4E75-935E-CDFA79A9E146}.Release|x64.ActiveCfg = Release|Any CPU
-		{D5EB155C-5365-4E75-935E-CDFA79A9E146}.Release|x64.Build.0 = Release|Any CPU
-		{D5EB155C-5365-4E75-935E-CDFA79A9E146}.Release|x86.ActiveCfg = Release|Any CPU
-		{D5EB155C-5365-4E75-935E-CDFA79A9E146}.Release|x86.Build.0 = Release|Any CPU
-		{D74C9C84-B359-49A5-8B0C-B296C7D8C3D3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D74C9C84-B359-49A5-8B0C-B296C7D8C3D3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D74C9C84-B359-49A5-8B0C-B296C7D8C3D3}.Debug|ARM64.ActiveCfg = Debug|Any CPU
-		{D74C9C84-B359-49A5-8B0C-B296C7D8C3D3}.Debug|ARM64.Build.0 = Debug|Any CPU
-		{D74C9C84-B359-49A5-8B0C-B296C7D8C3D3}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{D74C9C84-B359-49A5-8B0C-B296C7D8C3D3}.Debug|x64.Build.0 = Debug|Any CPU
-		{D74C9C84-B359-49A5-8B0C-B296C7D8C3D3}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{D74C9C84-B359-49A5-8B0C-B296C7D8C3D3}.Debug|x86.Build.0 = Debug|Any CPU
-		{D74C9C84-B359-49A5-8B0C-B296C7D8C3D3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D74C9C84-B359-49A5-8B0C-B296C7D8C3D3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D74C9C84-B359-49A5-8B0C-B296C7D8C3D3}.Release|ARM64.ActiveCfg = Release|Any CPU
-		{D74C9C84-B359-49A5-8B0C-B296C7D8C3D3}.Release|ARM64.Build.0 = Release|Any CPU
-		{D74C9C84-B359-49A5-8B0C-B296C7D8C3D3}.Release|x64.ActiveCfg = Release|Any CPU
-		{D74C9C84-B359-49A5-8B0C-B296C7D8C3D3}.Release|x64.Build.0 = Release|Any CPU
-		{D74C9C84-B359-49A5-8B0C-B296C7D8C3D3}.Release|x86.ActiveCfg = Release|Any CPU
-		{D74C9C84-B359-49A5-8B0C-B296C7D8C3D3}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
