+++ Raspite/Tags/CompoundTag.cs
@@ -1,0 +1,6 @@
+﻿namespace Raspite.Tags;
+
+public sealed class CompoundTag(Tag[] value, string name = "") : Tag<Tag[]>(value, name)
+{
+    public override byte Identifier => 10;
+}
