+++ Raspite/Tags/DoubleTag.cs
@@ -1,0 +1,6 @@
+﻿namespace Raspite.Tags;
+
+public sealed class DoubleTag(double value, string name = "") : Tag<double>(value, name)
+{
+    public override byte Identifier => 6;
+}
