+++ Raspite/Tags/FloatTag.cs
@@ -1,0 +1,6 @@
+﻿namespace Raspite.Tags;
+
+public sealed class FloatTag(float value, string name = "") : Tag<float>(value, name)
+{
+    public override byte Identifier => 5;
+}
