+++ Raspite/Tags/IntegerCollectionTag.cs
@@ -1,0 +1,6 @@
+﻿namespace Raspite.Tags;
+
+public sealed class IntegerCollectionTag(int[] value, string name = "") : Tag<int[]>(value, name)
+{
+    public override byte Identifier => 11;
+}
