+++ Raspite/Tags/IntegerTag.cs
@@ -1,0 +1,6 @@
+﻿namespace Raspite.Tags;
+
+public sealed class IntegerTag(int value, string name = "") : Tag<int>(value, name)
+{
+    public override byte Identifier => 3;
+}
