+++ Raspite/Tags/LongCollectionTag.cs
@@ -1,0 +1,6 @@
+﻿namespace Raspite.Tags;
+
+public sealed class LongCollectionTag(long[] value, string name = "") : Tag<long[]>(value, name)
+{
+    public override byte Identifier => 12;
+}
