+++ Raspite/Tags/LongTag.cs
@@ -1,0 +1,6 @@
+﻿namespace Raspite.Tags;
+
+public sealed class LongTag(long value, string name = "") : Tag<long>(value, name)
+{
+    public override byte Identifier => 4;
+}
