+++ Raspite/Tags/StringTag.cs
@@ -1,0 +1,6 @@
+﻿namespace Raspite.Tags;
+
+public sealed class StringTag(string value, string name = "") : Tag<string>(value, name)
+{
+    public override byte Identifier => 8;
+}
